✨ sort pipes by their names in usePipeList
