+++ src/pages/pipes/Pipes.tsx
@@ -1,28 +1,14 @@
 import type { ReactElement } from 'react';
-import { useQuery } from '@apollo/client';
 import { Link, generatePath } from 'react-router-dom';
 import { PIPE_PATH } from '../../navigation/paths';
-import useOrganization from '../../organization/useOrganization';
-import GET_PIPE_LIST_QUERY, {
-  GetPipeListData,
-  GetPipeListVariables
-} from './graphql/GET_PIPE_LIST_QUERY';
+import usePipeList from './hooks/usePipeList';
 
 function Pipes(): ReactElement {
-  const organization = useOrganization();
-
-  const { data } = useQuery<
-    GetPipeListData,
-    GetPipeListVariables
-  >(GET_PIPE_LIST_QUERY, {
-    variables: {
-      organizationId: organization.id,
-    }
-  });
+  const { pipes } = usePipeList();
 
   return (
     <ul>
-      {data?.organization.pipes.map((pipe) => (
+      {pipes.map((pipe) => (
         <li key={pipe.id}>
           <Link to={generatePath(PIPE_PATH, { pipeId: pipe.id })}>{pipe.name}</Link>
         </li>
